+++ SourceryRuntime/Sources/Method.swift
@@ -14,6 +14,9 @@ public final class MethodParameter: NSObject, SourceryModel, Typed, Annotated {
     /// Parameter type name
     public let typeName: TypeName
 
+    /// Parameter flag whether it's inout or not
+    public let `inout`: Bool
+
     // sourcery: skipEquality, skipDescription
     /// Parameter type, if known
     public var type: Type?
@@ -35,23 +38,25 @@ public final class MethodParameter: NSObject, SourceryModel, Typed, Annotated {
     public var __parserData: Any?
 
     /// :nodoc:
-    public init(argumentLabel: String?, name: String = "", typeName: TypeName, type: Type? = nil, defaultValue: String? = nil, annotations: [String: NSObject] = [:]) {
+    public init(argumentLabel: String?, name: String = "", typeName: TypeName, type: Type? = nil, defaultValue: String? = nil, annotations: [String: NSObject] = [:], `inout`: Bool = false) {
         self.typeName = typeName
         self.argumentLabel = argumentLabel
         self.name = name
         self.type = type
         self.defaultValue = defaultValue
         self.annotations = annotations
+        self.`inout` = `inout`
     }
 
     /// :nodoc:
-    public init(name: String = "", typeName: TypeName, type: Type? = nil, defaultValue: String? = nil, annotations: [String: NSObject] = [:]) {
+    public init(name: String = "", typeName: TypeName, type: Type? = nil, defaultValue: String? = nil, annotations: [String: NSObject] = [:], `inout`: Bool = false) {
         self.typeName = typeName
         self.argumentLabel = name
         self.name = name
         self.type = type
         self.defaultValue = defaultValue
         self.annotations = annotations
+        self.`inout` = `inout`
     }
 
     // sourcery:inline:MethodParameter.AutoCoding
@@ -62,6 +67,7 @@ public final class MethodParameter: NSObject, SourceryModel, Typed, Annotated {
             guard let typeName: TypeName = aDecoder.decode(forKey: "typeName") else { NSException.raise(NSExceptionName.parseErrorException, format: "Key '%@' not found.", arguments: getVaList(["typeName"])); fatalError() }; self.typeName = typeName
             self.type = aDecoder.decode(forKey: "type")
             self.defaultValue = aDecoder.decode(forKey: "defaultValue")
+            self.inout = aDecoder.decode(forKey: "inout")
             guard let annotations: [String: NSObject] = aDecoder.decode(forKey: "annotations") else { NSException.raise(NSExceptionName.parseErrorException, format: "Key '%@' not found.", arguments: getVaList(["annotations"])); fatalError() }; self.annotations = annotations
         }
 
@@ -73,6 +79,7 @@ public final class MethodParameter: NSObject, SourceryModel, Typed, Annotated {
             aCoder.encode(self.type, forKey: "type")
             aCoder.encode(self.defaultValue, forKey: "defaultValue")
             aCoder.encode(self.annotations, forKey: "annotations")
+            aCoder.encode(self.inout, forKey: "inout")
         }
         // sourcery:end
 }