+++ SourceryRuntime/Sources/TypeName.swift
@@ -66,6 +66,7 @@ public final class TypeName: NSObject, AutoCoding, AutoEquatable, AutoDiffable, 
             self.isOptional = false
         } else {
             name = name.bracketsBalancing()
+            name = name.trimmingPrefix("inout ").trimmingCharacters(in: .whitespaces)
             let isImplicitlyUnwrappedOptional = name.hasSuffix("!") || name.hasPrefix("ImplicitlyUnwrappedOptional<")
             let isOptional = name.hasSuffix("?") || name.hasPrefix("Optional<") || isImplicitlyUnwrappedOptional
             self.isImplicitlyUnwrappedOptional = isImplicitlyUnwrappedOptional